requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
docker-pycreds==0.4.0
filelock==3.14.0
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
pillow==10.3.0
platformdirs==4.2.2
protobuf==4.25.3
psutil==5.9.8
PyYAML==6.0.1
requests==2.32.2
sentry-sdk==2.3.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
sympy==1.12
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
wandb==0.17.0
